pyproject.toml
[project]
name = "clinia-doc-crawler"
version = "0.1.0"
description = "A Python project using uv and src layout."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "openai==1.59.6",
    "html2text==2024.2.26",
    "supabase==2.11.0",
    "Crawl4AI==0.4.247",
    "pydantic-ai",
    "logfire",
]

[dependency-groups]
dev = [
    "ruff==0.8.2",
    "pre-commit==3.7.1",
    "pytest>=8.3.3",
    "pytest-cov>=4.0.0",
    "ipykernel",
    "tiktoken",
    "nest_asyncio"
]

[tool.uv]
package=true
default-groups = ["dev"]

## Imports and formatting
[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = ["gen"]

[tool.ruff.lint]
ignore = ["C901", "E501", "F401", "F403", "F841", "W605"]
select = ["B", "B9", "C", "E", "F", "W", "I001"]

[tool.ruff.lint.mccabe]
max-complexity = 18


